+++ ckanext/toolbelt/cli/ckan/db.py
@@ -10,14 +10,21 @@ import ckan.plugins.toolkit as tk
 def db():
     pass
 
+
 @db.command()
 @click.option("-y", "--yes", is_flag=True, help="Confirm your intention.")
-@click.option("-k", "--keep", multiple=True, help="Tables that should not be cleaned")
+@click.option(
+    "-k", "--keep", multiple=True, help="Tables that should not be cleaned"
+)
 def clean(yes: bool, keep: tuple[str]):
     if not yes:
         tk.error_shout("This command will erase data from your portal's DB.")
-        tk.error_shout("All the datasets, organizations and users will be removed")
-        tk.error_shout("Run it with `--yes` flag if you know what you are doing.")
+        tk.error_shout(
+            "All the datasets, organizations and users will be removed"
+        )
+        tk.error_shout(
+            "Run it with `--yes` flag if you know what you are doing."
+        )
         raise click.Abort()
 
     model.repo.session.remove()
@@ -33,8 +40,8 @@ def clean(yes: bool, keep: tuple[str]):
             # skip `geometry_columns` or similar object from postgis
             continue
 
-        if table.name == 'alembic_version':
+        if table.name == "alembic_version":
             continue
         connection.execute('truncate "%s" cascade' % table.name)
     model.repo.session.commit()
-    click.secho('Database table data deleted', fg="green")
+    click.secho("Database table data deleted", fg="green")